+++ src/main/java/cc/landingzone/dreamweb/controller/LoginController.java
@@ -1,15 +1,11 @@
 package cc.landingzone.dreamweb.controller;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import javax.crypto.Mac;
-import javax.crypto.spec.SecretKeySpec;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -23,6 +19,7 @@ import cc.landingzone.dreamweb.service.ApiUserService;
 import cc.landingzone.dreamweb.service.UserService;
 import cc.landingzone.dreamweb.utils.HttpClientUtils;
 import cc.landingzone.dreamweb.utils.JsonUtils;
+import cc.landingzone.dreamweb.utils.SignatureUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -44,11 +41,6 @@ public class LoginController extends BaseController {
     private ApiUserService apiUserService;
 
     private static Logger logger = LoggerFactory.getLogger(LoginController.class);
-
-    private static final String DEFAULT_CHARSET = "UTF-8";
-    private static final long ONE_SECOND = 1000L;
-    private static final long ONE_MINUTE = 60 * ONE_SECOND;
-    private static final long THIRTY_MINUTES = 30 * ONE_MINUTE;
 
     @RequestMapping("/")
     public void hello(HttpServletRequest request, HttpServletResponse response) {
@@ -99,9 +91,9 @@ public class LoginController extends BaseController {
                 .collect(Collectors.joining("&"));
 
             // 校验签名
-            checkSignature(params, apiUser.getAccessKeySecret(), signature);
+            SignatureUtils.checkSignature(params, apiUser.getAccessKeySecret(), signature);
             // 检查签名是否过期
-            checkTimestamp(now, timestamp);
+            SignatureUtils.checkTimestamp(now, timestamp);
 
             // 查询用户信息
             User user = userService.getUserByLoginName(loginName);
@@ -193,85 +185,5 @@ public class LoginController extends BaseController {
             result = e.getMessage();
         }
         outputToString(response, result);
-    }
-
-    private static void checkSignature(String params, String accessKeySecret, String signature) {
-        Assert.isTrue(signature.equals(generateSignature(params, accessKeySecret)), "签名验证失败!");
-    }
-
-    private static String generateSignature(String params, String accessKeySecret) {
-        String stringToSign = params;
-        try {
-            Mac mac = Mac.getInstance("HmacSHA1");
-            SecretKeySpec signingKey = new SecretKeySpec(accessKeySecret.getBytes(DEFAULT_CHARSET), mac.getAlgorithm());
-            mac.init(signingKey);
-            byte[] rawHmac = mac.doFinal(stringToSign.getBytes(DEFAULT_CHARSET));
-            return toHexString(rawHmac);
-        } catch (Exception e) {
-            logger.error(e.getMessage(), e);
-        }
-        return null;
-    }
-
-    /**
-     * 把二进制转化为小写的十六进制
-     *
-     * @param bytes
-     * @return
-     */
-    private static String toHexString(final byte[] bytes) {
-        StringBuilder hexString = new StringBuilder();
-        for (int i = 0; i < bytes.length; i++) {
-            String hex = Integer.toHexString(bytes[i] & 0xFF);
-            if (hex.length() == 1) {
-                hexString.append("0");
-            }
-            hexString.append(hex);
-        }
-        return hexString.toString();
-    }
-
-    private static void checkTimestamp(long now, String timestamp) {
-        Assert.isTrue(Math.abs(now - Long.parseLong(timestamp)) <= THIRTY_MINUTES, "签名已经过期!");
-    }
-
-    /**
-     * 生成API登录需要的参数
-     */
-    private static void generateApiLoginParams() {
-        // 创建参数表
-        Map<String, String> params = new HashMap<>();
-        // Access Key ID
-        params.put("accessKeyId", CommonConstants.TEST_API_ACCESS_KEY_ID);
-        // 13位时间戳
-        params.put("timestamp", String.valueOf(System.currentTimeMillis()));
-        // 自定义参数
-        params.put("loginName", CommonConstants.TEST_LOGIN_NAME);
-
-        // 按照参数名称的字典顺序对请求中所有参数(不包括signature参数本身)进行排序
-        String[] keyArray = params.keySet().toArray(new String[0]);
-        Arrays.sort(keyArray);
-
-        // 将参数名称和值用"="进行连接,得到形如"key=value"的字符串
-        // 将"="连接得到的参数组合按顺序依次用"&"进行连接,得到形如"key1=value1&key2=value2..."的字符串
-        StringBuilder stringBuilder = new StringBuilder();
-        boolean isFirst = true;
-        for (String key : keyArray) {
-            if (!isFirst) {
-                stringBuilder.append("&");
-            } else {
-                isFirst = false;
-            }
-            stringBuilder.append(key).append("=").append(params.get(key));
-        }
-        String needSignature = stringBuilder.toString();
-
-        // 生成签名
-        String signature = generateSignature(needSignature, CommonConstants.TEST_API_ACCESS_KEY_SECRET);
-        System.out.println(needSignature + "&signature=" + signature);
-    }
-
-    public static void main(String[] args) {
-        generateApiLoginParams();
     }
 }
