+++ src/main/resources/application.properties
@@ -46,9 +46,3 @@ dreamweb.weixin.landingzone.id=<your_weixin_keyid>
 dreamweb.weixin.landingzone.secret=<your_weixin_keysecret>
 # API登出成功后跳转的URL(一般为客户平台的地址)
 dreamweb.api_logout_success_url=<your_api_logout_success_url>
-
-# dreamweb 测试用配置
-dreamweb.test.api_accesskeyid=<test_api_accesskeyid>
-dreamweb.test.api_accesskeysecret=<test_api_accesskeysecret>
-dreamweb.test.login_name=<test_login_name>
-
